refactor: remove secretSquarePoll schema
